requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dialogflow==0.5.2
Flask==1.0.2
google-api-core==1.8.1
google-auth==1.6.3
googleapis-common-protos==1.5.8
grpcio==1.19.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
protobuf==3.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyJWT==1.7.1
PySocks==1.6.8
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
soupsieve==1.8
twilio==6.25.2
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.15.1
